Change provider params from int to size_t

+++ providers/common/include/internal/ciphers/cipher_ccm.h
@@ -36,7 +36,7 @@ typedef struct prov_ccm_st {
     size_t l, m;                /* L and M parameters from RFC3610 */
     size_t keylen;
     int tls_aad_len;            /* TLS AAD length */
-    int tls_aad_pad_sz;
+    size_t tls_aad_pad_sz;
     unsigned char iv[AES_BLOCK_SIZE];
     unsigned char buf[AES_BLOCK_SIZE];
     CCM128_CONTEXT ccm_ctx;
